ANSI C disallows unwidened types in va_start.

+++ include/mqtt.h
@@ -801,7 +801,7 @@ ssize_t mqtt_pack_pubxxx_request(uint8_t *buf, size_t bufsz,
  *          packet, a negative value if there was a protocol violation.
  */
 ssize_t mqtt_pack_subscribe_request(uint8_t *buf, size_t bufsz, 
-                                    uint16_t packet_id, 
+                                    unsigned int packet_id, 
                                     ...); /* null terminated */
 
 /** 
@@ -835,7 +835,7 @@ ssize_t mqtt_pack_subscribe_request(uint8_t *buf, size_t bufsz,
  *          packet, a negative value if there was a protocol violation.
  */
 ssize_t mqtt_pack_unsubscribe_request(uint8_t *buf, size_t bufsz, 
-                                      uint16_t packet_id, 
+                                      unsigned int packet_id, 
                                       ...); /* null terminated */
 
 /**
